Component tests for BWFilter and SepiaFilter.
This adds component tests for BWFilter and SepiaFilter, for both the GD
and Imagick	backends. These tests execute the conversion and compare the
results to reference images, rather than just mocking out the relevant
backend function calls.

Writing these tests exposed an additional weakness, namely that GD and
Imagick output quite different images. For now, I have made this
apparent by introducing a pair of reference images. It would be nice to
alter the implementation of one of the backends so that the output was
more consistent across both.